Data serialization and compression work for future save & load system.

+++ Assets/Scripts/_New/Terrain/Block.cs
@@ -4,7 +4,8 @@ using UnityEngine;
 
 namespace UnityCommunityVoxelProject.Terrain
 {
-    public enum Block : uint
+    
+    public enum Block : byte
     {
         Air, 
         Dirt,
